mention cookie requirement, resolves #168
The main page displays a error box “your browser is not supported or JavaScript
is disabled” even though the root cause is that Cookies are disabled. This is
very confusing. This change mentions Cookies in the message as well. Evidently,
the best solution would be not to require cookies at all.